Refactor: Extract styled components to separate file
